# VideoGameApiWithKafka/Video Game API 1.9/openapi.yaml
openapi: 3.0.0
info:
  title: Video Game Exchange
  version: 1.0.0
servers:
- url: http://localhost:8080
paths:
...
  /offers:
    get:
      operationId: listOffers
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Offer'
                type: array
          description: A paged array of Offers
      summary: List All Offers
      tags:
      - offers
      x-accepts: application/json
      x-tags:
      - tag: offers
    post:
      operationId: createOffer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Offer'
        required: true
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Create an Offer
      tags:
      - offers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: offers
  /offers/{offerId}:
    delete:
      operationId: deleteOffer
      parameters:
      - description: The id of the offer to delete
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Delete one Offer
      tags:
      - offers
      x-accepts: application/json
      x-tags:
      - tag: offers
    get:
      operationId: getOneOffer
      parameters:
      - description: The id of the offer to get
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Offer'
          description: One Offer Object
      summary: Get One Offer
      tags:
      - offers
      x-accepts: application/json
      x-tags:
      - tag: offers
    patch:
      operationId: updateSomeOffer
      parameters:
      - description: The id of the offer to update
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Offer'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update Some offer Attributes
      tags:
      - offers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: offers
    put:
      operationId: updateOffer
      parameters:
      - description: The id of the offer to update
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Offer'
        required: true
      responses:
        "204":
          description: Null response
        default:
          description: unexpected error
      summary: Update all Offer Attributes
      tags:
      - offers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: offers
  /offers/{offerId}/{gameId}/{isSent}:
    post: 
      parameters:
      - description: The id of the offer to add a Game to
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the game to add to offer
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      - description: boolean to determine if game is being sent or not
        explode: false
        in: path
        name: isSent
        required: true
        schema:
          type: boolean
        style: simple
      operationId: addGameOffer
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Add a game to a offer
      tags:
      - users
      - games
      - offers
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
    delete: 
      parameters:
      - description: The id of the offer to delete a Game from
        explode: false
        in: path
        name: offerId
        required: true
        schema:
          type: string
        style: simple
      - description: The id of the game to delete from offer
        explode: false
        in: path
        name: gameId
        required: true
        schema:
          type: string
        style: simple
      - description: boolean to determine if game is being sent or not
        explode: false
        in: path
        name: isSent
        required: true
        schema:
          type: boolean
        style: simple
      operationId: deleteGameOffer
      responses:
        "201":
          description: Null response
        default:
          description: unexpected error
      summary: Remove a Game from a User  
      tags:
      - users
      - games
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: users
components:
  schemas:
...
    Offer:
      example:
        toUser:
        - null
        - null
        sentGames:
        - null
        - null
        fromUser:
        - null
        - null
        receivedGames:
        - null
        - null
        timeCreated: 2000-01-23T04:56:07.000+00:00
        id: 0
        state: state
      properties:
        id:
          format: int64
          type: integer
        fromUser:
          type: string
        sentGames:
          items:
            type: string
          type: array
        receivedGames:
          items:
            type: string
          type: array
        toUser:
          type: string
        timeCreated:
          type: string
        state:
          type: string
      required:
      - fromUser
      - id
      - receivedGames
      - sentGames
      - state
      - timeCreated
      - toUser
      type: object
...